+++ src/BalancelleBundle/Controller/StructureController.php
@@ -4,11 +4,13 @@ namespace BalancelleBundle\Controller;
 
 use BalancelleBundle\Entity\Structure;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+//use Symfony\Bundle\MakerBundle\Str;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use BalancelleBundle\Form\StructureType;
+use BalancelleBundle\Form\MailType;
 
 /**
  * Structure controller.
@@ -146,5 +148,49 @@ class StructureController extends Controller implements FamilleInterface
             ->setMethod('DELETE')
             ->getForm()
         ;
+    }
+
+    /**
+     * Envoie un mail aux parents de la structure
+     * @param Request $request
+     * @param Structure $structure - la structure concernée
+     * @return Response
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
+     */
+    public function envoyerMailAction(Request $request, Structure $structure)
+    {
+        $form = $this->createForm(
+            MailType::class
+        );
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $documents = $form->getData()['documents'];
+            array_pop($documents);
+            $this->get('communication')->envoyerMailStructure(
+                $structure,
+                $form->getData()['sujet'],
+                $form->getData()['message'],
+                $documents
+            );
+            $succes = 'Votre email a bien été envoyé aux parents';
+            $this->addFlash('success', $succes);
+            return $this->redirectToRoute(
+                'structure_edit',
+                array('id' => $structure->getId())
+            );
+        }
+        $titre = 'Envoyer un email aux parents de la structure ';
+        $titre .= $structure->getNom();
+        return $this->render(
+            '@Balancelle/Communication/email_create.html.twig',
+            array(
+                'structure' => $structure,
+                'form' => $form->createView(),
+                'titre' => $titre
+            )
+        );
     }
 }