+++ src/BalancelleBundle/Form/MailType.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace BalancelleBundle\Form;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CollectionType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use FOS\CKEditorBundle\Form\Type\CKEditorType;
+
+
+class MailType extends AbstractType
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('sujet', TextType::class, ['required' => true])
+            //->add('message', TextareaType::class, ['required' => true])
+            ->add('message', CKEditorType::class, array(
+                /*'config' => array(
+                    'toolbar' => 'balancelle'
+                ),*/
+                'config_name' => 'standard'
+            ))
+            ->add('documents', CollectionType::class, array(
+                'entry_type'   		=> DocumentType::class,
+                'prototype'			=> true,
+                'allow_add'			=> true,
+                'allow_delete'		=> true,
+                'required'			=> false,
+                'label'			=> false,
+            ));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults(array());
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getBlockPrefix()
+    {
+        return 'balancellebundle_structure';
+    }
+
+}