+++ changelog.md
@@ -6,6 +6,12 @@ tristan.fleury@labalancelle.yo.fr
 **Site intranet de l'association La Balancelle**    
 
 ------------
+## [10/11/2019] [v0.0.2][task53]
+    - mise en place envoie de mail par structure
+    - gestion des pièces jointes
+    - installation ckeditor: sudo apt-get install php7.2-zip
+    - ajout du ckeditor
+    
 ## [08/11/2019] [v0.0.2][bug58]
     - correction du bug: séparation de l'instanciation des dates matin et am
         pour la génération des permanences (le setTime de l'am écrasait
@@ -17,7 +23,7 @@ tristan.fleury@labalancelle.yo.fr
     
 ## [30/10/2019] [v0.0.2][bug52]
     - réinitialisation de la variable to (array) à chaque permanence ciblée
-    
+       
 ## [28/10/2019] [v0.0.2][bug56]
     - correction déconnexion impossible
     - désactivation des assests chart